+++ packages/protocol/test/tokenvault/ERC20Vault.t.sol
@@ -5,7 +5,6 @@ import { AddressManager } from "../../contracts/common/AddressManager.sol";
 import { AddressResolver } from "../../contracts/common/AddressResolver.sol";
 import { Bridge } from "../../contracts/bridge/Bridge.sol";
 import { BridgedERC20 } from "../../contracts/tokenvault/BridgedERC20.sol";
-import { BridgeErrors } from "../../contracts/bridge/BridgeErrors.sol";
 import { FreeMintERC20 } from "../../contracts/test/erc20/FreeMintERC20.sol";
 import { SignalService } from "../../contracts/signal/SignalService.sol";
 import { TaikoToken } from "../../contracts/L1/TaikoToken.sol";