+++ packages/protocol/test/tokenvault/ERC721Vault.t.sol
@@ -11,14 +11,10 @@ import {
 } from "../TestBase.sol";
 import { AddressManager } from "../../contracts/common/AddressManager.sol";
 import { IBridge, Bridge } from "../../contracts/bridge/Bridge.sol";
-import { BridgeData } from "../../contracts/bridge/BridgeData.sol";
-import { BridgeErrors } from "../../contracts/bridge/BridgeErrors.sol";
 import { BaseNFTVault } from "../../contracts/tokenvault/BaseNFTVault.sol";
 import { ERC721Vault } from "../../contracts/tokenvault/ERC721Vault.sol";
 import { BridgedERC721 } from "../../contracts/tokenvault/BridgedERC721.sol";
 import { EtherVault } from "../../contracts/bridge/EtherVault.sol";
-import { LibBridgeStatus } from
-    "../../contracts/bridge/libs/LibBridgeStatus.sol";
 import { SignalService } from "../../contracts/signal/SignalService.sol";
 import { ICrossChainSync } from "../../contracts/common/ICrossChainSync.sol";
 import { ERC721 } from "@openzeppelin/contracts/token/ERC721/ERC721.sol";
@@ -77,7 +73,7 @@ contract PrankDestBridge {
         destERC721Vault = ERC721Vault(addr);
     }
 
-    function sendMessage(BridgeData.Message memory message)
+    function sendMessage(IBridge.Message memory message)
         external
         payable
         returns (bytes32 msgHash)
@@ -639,7 +635,7 @@ contract ERC721VaultTest is TestBase {
         // srcChain, because we mock the bridge functions, but good to have data
         // here so that it could have been hashed back to the exact same bytes32
         // value - if we were not mocking.
-        BridgeData.Message memory message;
+        IBridge.Message memory message;
         message.srcChainId = 31_337;
         message.destChainId = destChainId;
         message.user = Alice;
@@ -1020,7 +1016,7 @@ contract ERC721VaultTest is TestBase {
 
     function _recallMessage(
         Bridge srcBridge,
-        BridgeData.Message memory message,
+        IBridge.Message memory message,
         bytes memory proof
     )
         internal